requirements.txt
streamlit
requests
pandas
python-docx
phpserialize
